+++ src/class-ss-integrations.php
@@ -35,7 +35,6 @@ class Integrations {
 			'cookieyes'     => CookieYes_Integration::class,
 			'brizy'         => Brizy_Integration::class,
 			'complianz'     => Complianz_Integration::class,
-			'brizy'         => Brizy_Integration::class,
 			'jetpack'       => Jetpack_Integration::class,
 			'multilingual'  => Multilingual_Integration::class,
 			'github'        => Github_Integration::class,
@@ -68,7 +67,7 @@ class Integrations {
 		require_once $path . 'pro/class-github-integration.php';
 		require_once $path . 'pro/class-multilingual-integration.php';
 		require_once $path . 'pro/class-shortpixel-integration.php';
-		require_once $path . 'class-complianz-integration.php';
+		require_once $path . 'pro/class-complianz-integration.php';
 
 		// SimplyCDN.
 		require_once $path . 'simply-cdn/class-simply-cdn-integration.php';
